requirements.txt
# ============================================================
# FILE: requirements.txt
# ============================================================
pyrogram>=2.0.0
tgcrypto>=1.2.5
python-dotenv>=1.0.0